Switch the generated thumnails from GIF to PNG to avoid paleness and thumbnail overlap
The ppmquantall program used to limit the number of colours used
across the thumbnails caused the thumbnails of one image to overlap
with the next

It looked like the top of the next image ended up as
the bottom of the current.

It's possible that ppmquantall put all images into one big image and
then split it up again after limiting the number of colours, and that
there was an alignment problem in the split?

Also the quantized GIF thumbnails became so pale it was hard to see
what was one them.

But unlike in 1996, PNGs are universally supported, and GIF is just an
unpleasant legacy, so the switch makes sense.